requirements.txt
aniso8601==7.0.0
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.0.3
Flask-Cors==3.0.8
Flask-RESTful==0.3.7
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
MarkupSafe==1.1.1
PyMySQL==0.9.3
pyrsistent==0.15.4
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
six==1.12.0
swagger-ui-py==0.1.9
urllib3==1.25.3
uuid==1.30
validator-collection==1.3.5
Werkzeug==0.15.4
